policy gradient training pipeline
